+++ adl/src/cli/pr_cmd.rs
@@ -6,6 +6,19 @@ use std::os::unix::fs as unix_fs;
 use std::path::{Path, PathBuf};
 use std::process::Command;
 
+use super::pr_cmd_args::{
+    parse_finish_args, parse_init_args, parse_preflight_args, parse_ready_args, parse_start_args,
+};
+#[cfg(test)]
+use super::pr_cmd_prompt::{
+    infer_required_outcome_type, infer_wp_from_title, load_issue_prompt,
+    parse_issue_number_from_url, resolve_issue_body, version_from_labels_csv,
+};
+use super::pr_cmd_prompt::{
+    normalize_labels_csv, render_generated_issue_prompt, resolve_issue_prompt_path,
+    resolve_issue_scope_and_slug_from_local_state, validate_issue_prompt_exists,
+    version_from_title,
+};
 use ::adl::control_plane::{
     card_input_path, card_output_path, resolve_cards_root, resolve_primary_checkout_root,
     sanitize_slug, IssueRef,
@@ -13,74 +26,6 @@ use ::adl::control_plane::{
 
 const DEFAULT_VERSION: &str = "v0.86";
 const DEFAULT_NEW_LABELS: &str = "track:roadmap,type:task,area:tools";
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct InitArgs {
-    issue: u32,
-    slug: Option<String>,
-    title_arg: Option<String>,
-    no_fetch_issue: bool,
-    version: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct StartArgs {
-    issue: u32,
-    prefix: String,
-    slug: Option<String>,
-    title_arg: Option<String>,
-    no_fetch_issue: bool,
-    version: Option<String>,
-    allow_open_pr_wave: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct ReadyArgs {
-    issue: u32,
-    slug: Option<String>,
-    version: Option<String>,
-    no_fetch_issue: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct PreflightArgs {
-    issue: u32,
-    version: Option<String>,
-    slug: Option<String>,
-    no_fetch_issue: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct FinishArgs {
-    issue: u32,
-    title: String,
-    extra_body: Option<String>,
-    paths: String,
-    no_checks: bool,
-    no_close: bool,
-    ready: bool,
-    allow_gitignore: bool,
-    input_path: Option<PathBuf>,
-    output_path: Option<PathBuf>,
-    no_open: bool,
-    merge_mode: bool,
-    idempotent: bool,
-}
-
-#[cfg(test)]
-#[derive(Debug, Deserialize)]
-struct IssuePromptFrontMatter {
-    title: String,
-    labels: Vec<String>,
-    issue_number: u32,
-}
-
-#[cfg(test)]
-#[derive(Debug)]
-struct IssuePromptDoc {
-    front_matter: IssuePromptFrontMatter,
-    body: String,
-}
 
 #[derive(Debug, Clone, PartialEq, Eq, Deserialize)]
 struct OpenPullRequest {
@@ -644,261 +589,6 @@ fn real_pr_init(args: &[String]) -> Result<()> {
     println!("  STATE    ISSUE_AND_BUNDLE_READY");
     eprintln!("• Done.");
     Ok(())
-}
-
-fn parse_init_args(args: &[String]) -> Result<InitArgs> {
-    let issue_raw = args
-        .first()
-        .ok_or_else(|| anyhow!("init: missing <issue> number"))?;
-    let issue = issue_raw
-        .parse::<u32>()
-        .with_context(|| format!("invalid issue number: {issue_raw}"))?;
-    let mut parsed = InitArgs {
-        issue,
-        slug: None,
-        title_arg: None,
-        no_fetch_issue: false,
-        version: None,
-    };
-
-    let mut i = 1;
-    while i < args.len() {
-        match args[i].as_str() {
-            "--slug" => {
-                parsed.slug = Some(require_value(args, i, "init", "--slug")?);
-                i += 1;
-            }
-            "--title" => {
-                parsed.title_arg = Some(require_value(args, i, "init", "--title")?);
-                i += 1;
-            }
-            "--no-fetch-issue" => parsed.no_fetch_issue = true,
-            "--version" => {
-                parsed.version = Some(require_value(args, i, "init", "--version")?);
-                i += 1;
-            }
-            other => bail!("init: unknown arg: {other}"),
-        }
-        i += 1;
-    }
-    Ok(parsed)
-}
-
-fn parse_start_args(args: &[String]) -> Result<StartArgs> {
-    let issue_raw = args
-        .first()
-        .ok_or_else(|| anyhow!("start: missing <issue> number"))?;
-    let issue = issue_raw
-        .parse::<u32>()
-        .with_context(|| format!("invalid issue number: {issue_raw}"))?;
-    let mut parsed = StartArgs {
-        issue,
-        prefix: "codex".to_string(),
-        slug: None,
-        title_arg: None,
-        no_fetch_issue: false,
-        version: None,
-        allow_open_pr_wave: false,
-    };
-    let mut i = 1;
-    while i < args.len() {
-        match args[i].as_str() {
-            "--prefix" => {
-                parsed.prefix = require_value(args, i, "start", "--prefix")?;
-                i += 1;
-            }
-            "--slug" => {
-                parsed.slug = Some(require_value(args, i, "start", "--slug")?);
-                i += 1;
-            }
-            "--title" => {
-                parsed.title_arg = Some(require_value(args, i, "start", "--title")?);
-                i += 1;
-            }
-            "--version" => {
-                parsed.version = Some(require_value(args, i, "start", "--version")?);
-                i += 1;
-            }
-            "--no-fetch-issue" => parsed.no_fetch_issue = true,
-            "--allow-open-pr-wave" => parsed.allow_open_pr_wave = true,
-            other => bail!("start: unknown arg: {other}"),
-        }
-        i += 1;
-    }
-    Ok(parsed)
-}
-
-fn parse_preflight_args(args: &[String]) -> Result<PreflightArgs> {
-    let issue_raw = args
-        .first()
-        .ok_or_else(|| anyhow!("preflight: missing <issue> number"))?;
-    let issue = issue_raw
-        .parse::<u32>()
-        .with_context(|| format!("invalid issue number: {issue_raw}"))?;
-    let mut parsed = PreflightArgs {
-        issue,
-        version: None,
-        slug: None,
-        no_fetch_issue: false,
-    };
-    let mut i = 1;
-    while i < args.len() {
-        match args[i].as_str() {
-            "--slug" => {
-                parsed.slug = Some(require_value(args, i, "preflight", "--slug")?);
-                i += 1;
-            }
-            "--version" => {
-                parsed.version = Some(require_value(args, i, "preflight", "--version")?);
-                i += 1;
-            }
-            "--no-fetch-issue" => parsed.no_fetch_issue = true,
-            other => bail!("preflight: unknown arg: {other}"),
-        }
-        i += 1;
-    }
-    Ok(parsed)
-}
-
-fn parse_ready_args(args: &[String]) -> Result<ReadyArgs> {
-    let issue_raw = args
-        .first()
-        .ok_or_else(|| anyhow!("ready: missing <issue> number"))?;
-    let issue = issue_raw
-        .parse::<u32>()
-        .with_context(|| format!("invalid issue number: {issue_raw}"))?;
-    let mut parsed = ReadyArgs {
-        issue,
-        slug: None,
-        version: None,
-        no_fetch_issue: false,
-    };
-    let mut i = 1;
-    while i < args.len() {
-        match args[i].as_str() {
-            "--slug" => {
-                parsed.slug = Some(require_value(args, i, "ready", "--slug")?);
-                i += 1;
-            }
-            "--version" => {
-                parsed.version = Some(require_value(args, i, "ready", "--version")?);
-                i += 1;
-            }
-            "--no-fetch-issue" => parsed.no_fetch_issue = true,
-            other => bail!("ready: unknown arg: {other}"),
-        }
-        i += 1;
-    }
-    Ok(parsed)
-}
-
-fn parse_finish_args(args: &[String]) -> Result<FinishArgs> {
-    let issue = args
-        .first()
-        .ok_or_else(|| anyhow!("finish: missing <issue> number"))?;
-    let issue = issue
-        .parse::<u32>()
-        .with_context(|| format!("finish: invalid issue number: {issue}"))?;
-
-    let mut parsed = FinishArgs {
-        issue,
-        title: String::new(),
-        extra_body: None,
-        paths: ".".to_string(),
-        no_checks: false,
-        no_close: false,
-        ready: false,
-        allow_gitignore: false,
-        input_path: None,
-        output_path: None,
-        no_open: false,
-        merge_mode: false,
-        idempotent: false,
-    };
-
-    let mut i = 1;
-    while i < args.len() {
-        match args[i].as_str() {
-            "--title" => {
-                parsed.title = require_value(args, i, "finish", "--title")?;
-                i += 2;
-            }
-            "--body" => {
-                parsed.extra_body = Some(require_value(args, i, "finish", "--body")?);
-                i += 2;
-            }
-            "--paths" => {
-                parsed.paths = require_value(args, i, "finish", "--paths")?;
-                i += 2;
-            }
-            "--no-checks" => {
-                parsed.no_checks = true;
-                i += 1;
-            }
-            "--no-close" => {
-                parsed.no_close = true;
-                i += 1;
-            }
-            "--ready" => {
-                parsed.ready = true;
-                i += 1;
-            }
-            "--allow-gitignore" => {
-                parsed.allow_gitignore = true;
-                i += 1;
-            }
-            "-f" | "--file" | "--input" => {
-                parsed.input_path = Some(PathBuf::from(require_value(
-                    args,
-                    i,
-                    "finish",
-                    args[i].as_str(),
-                )?));
-                i += 2;
-            }
-            "--output" | "--output-card" | "--output-card-file" => {
-                parsed.output_path = Some(PathBuf::from(require_value(
-                    args,
-                    i,
-                    "finish",
-                    args[i].as_str(),
-                )?));
-                i += 2;
-            }
-            "--no-open" => {
-                parsed.no_open = true;
-                i += 1;
-            }
-            "--open" => {
-                parsed.no_open = false;
-                i += 1;
-            }
-            "--merge" | "--auto-merge" => {
-                parsed.merge_mode = true;
-                i += 1;
-            }
-            "--idempotent" => {
-                parsed.idempotent = true;
-                i += 1;
-            }
-            other => bail!("finish: unknown arg: {other}"),
-        }
-    }
-
-    if parsed.title.trim().is_empty() {
-        bail!("finish: --title is required");
-    }
-    if parsed.merge_mode && parsed.no_checks {
-        bail!("finish: --merge requires checks; remove --no-checks");
-    }
-
-    Ok(parsed)
-}
-
-fn require_value(args: &[String], index: usize, cmd: &str, flag: &str) -> Result<String> {
-    args.get(index + 1)
-        .cloned()
-        .ok_or_else(|| anyhow!("{cmd}: {flag} requires a value"))
 }
 
 fn repo_root() -> Result<PathBuf> {
@@ -1918,38 +1608,6 @@ fn branch_matches_started_state(recorded: &str, actual_branch: &str) -> bool {
     recorded.starts_with("TBD (run pr.sh start ")
 }
 
-fn resolve_issue_scope_and_slug_from_local_state(
-    repo_root: &Path,
-    issue: u32,
-) -> Result<Option<(String, String)>> {
-    let issue_dir = format!("issue-{:04}__", issue);
-    let adl_root = repo_root.join(".adl");
-    if !adl_root.is_dir() {
-        return Ok(None);
-    }
-    let mut matches = Vec::new();
-    for scope_entry in fs::read_dir(&adl_root)? {
-        let scope_entry = scope_entry?;
-        let tasks = scope_entry.path().join("tasks");
-        if !tasks.is_dir() {
-            continue;
-        }
-        for entry in fs::read_dir(&tasks)? {
-            let entry = entry?;
-            let name = entry.file_name();
-            let name = name.to_string_lossy();
-            if let Some(slug) = name.strip_prefix(&issue_dir) {
-                matches.push((
-                    scope_entry.file_name().to_string_lossy().to_string(),
-                    slug.to_string(),
-                ));
-            }
-        }
-    }
-    matches.sort();
-    Ok(matches.into_iter().next())
-}
-
 fn ensure_source_issue_prompt(
     repo_root: &Path,
     repo: &str,
@@ -2010,155 +1668,6 @@ fn ensure_source_issue_prompt(
     }
     fs::write(&source_path, content)?;
     Ok(source_path)
-}
-
-fn render_generated_issue_prompt(
-    issue: u32,
-    slug: &str,
-    title: &str,
-    labels_csv: &str,
-    issue_url: &str,
-) -> String {
-    let wp = infer_wp_from_title(title);
-    let outcome_type = infer_required_outcome_type(labels_csv, title);
-    let label_lines = labels_csv
-        .split(',')
-        .map(str::trim)
-        .filter(|label| !label.is_empty())
-        .map(|label| format!("  - \"{label}\""))
-        .collect::<Vec<_>>()
-        .join("\n");
-
-    format!(
-        "---\nissue_card_schema: adl.issue.v1\nwp: \"{wp}\"\nslug: \"{slug}\"\ntitle: \"{title}\"\nlabels:\n{label_lines}\nissue_number: {issue}\nstatus: \"draft\"\naction: \"edit\"\ndepends_on: []\nmilestone_sprint: \"Pending sprint assignment\"\nrequired_outcome_type:\n  - \"{outcome_type}\"\nrepo_inputs: []\ncanonical_files: []\ndemo_required: false\ndemo_names: []\nissue_graph_notes:\n  - \"Bootstrap-generated from GitHub issue metadata because no canonical local issue prompt existed yet.\"\npr_start:\n  enabled: true\n  slug: \"{slug}\"\n---\n\n# {title}\n\n## Summary\n\nBootstrap-generated local source prompt for issue #{issue}.\n\n## Goal\n\nTranslate the GitHub issue into the canonical local STP/task-bundle flow and refine this prompt before execution as needed.\n\n## Required Outcome\n\nThis issue currently defaults to a required outcome type of `{outcome_type}`. Refine this if the issue requires a different outcome or a combination.\n\n## Deliverables\n\n- one bounded, reviewable outcome matching the issue scope\n- updated canonical docs/code/tests/demo artifacts as required by the issue\n\n## Acceptance Criteria\n\n- the issue title and labels are reflected in the local source prompt\n- the task can proceed through `pr init`, `pr start`, and card editing without manual bootstrap repair\n\n## Repo Inputs\n\n- {issue_url}\n\n## Dependencies\n\n- none recorded yet\n\n## Demo Expectations\n\n- No demo is required by default. Update this section if the issue requires a proof surface.\n\n## Non-goals\n\n- changing milestone scope without recording it explicitly\n- ad-hoc local workflow drift outside the tracked issue flow\n\n## Issue-Graph Notes\n\n- This prompt was generated automatically because the canonical local issue prompt was missing.\n- Review and refine it before substantive implementation work begins.\n\n## Notes\n\n- GitHub issue: {issue_url}\n\n## Tooling Notes\n\n- Generated by `pr.sh` bootstrap fallback.\n"
-    )
-}
-
-fn infer_wp_from_title(title: &str) -> String {
-    if let Some(start) = title.find("[WP-") {
-        if let Some(end_rel) = title[start + 1..].find(']') {
-            return title[start + 1..start + 1 + end_rel].to_string();
-        }
-    }
-    "unassigned".to_string()
-}
-
-fn infer_required_outcome_type(labels_csv: &str, title: &str) -> &'static str {
-    let lowered = format!("{} {}", labels_csv.to_lowercase(), title.to_lowercase());
-    if lowered.contains("type:docs")
-        || lowered.contains("area:docs")
-        || lowered.contains("[docs]")
-        || lowered.contains("type:design")
-    {
-        return "docs";
-    }
-    if lowered.contains("type:test") || lowered.contains("area:tests") || lowered.contains("[test]")
-    {
-        return "tests";
-    }
-    if lowered.contains("area:demo") || lowered.contains("[demo]") {
-        return "demo";
-    }
-    "code"
-}
-
-#[cfg(test)]
-fn version_from_labels_csv(labels_csv: &str) -> Option<String> {
-    labels_csv
-        .split(',')
-        .map(str::trim)
-        .find_map(|label| label.strip_prefix("version:").map(str::to_string))
-}
-
-fn version_from_title(title: &str) -> Option<String> {
-    let start = title.find("[v")?;
-    let rest = &title[start + 1..];
-    let end = rest.find(']')?;
-    Some(rest[..end].to_string())
-}
-
-fn validate_issue_prompt_exists(path: &Path) -> Result<()> {
-    if !path.is_file() {
-        bail!("missing canonical source issue prompt: {}", path.display());
-    }
-    Ok(())
-}
-
-fn resolve_issue_prompt_path(repo_root: &Path, issue_ref: &IssueRef) -> Result<PathBuf> {
-    let preferred = issue_ref.issue_prompt_path(repo_root);
-    if preferred.is_file() {
-        return Ok(preferred);
-    }
-
-    let legacy = issue_ref.legacy_issue_prompt_path(repo_root);
-    if legacy.is_file() {
-        return Ok(legacy);
-    }
-
-    bail!(
-        "missing canonical source issue prompt: {}",
-        preferred.display()
-    )
-}
-
-#[cfg(test)]
-fn load_issue_prompt(path: &Path) -> Result<IssuePromptDoc> {
-    let text = fs::read_to_string(path)
-        .with_context(|| format!("failed to read issue prompt '{}'", path.display()))?;
-    let mut parts = text.splitn(3, "---");
-    let _ = parts.next();
-    let front_matter = parts
-        .next()
-        .ok_or_else(|| anyhow!("missing front matter in '{}'", path.display()))?;
-    let body = parts
-        .next()
-        .ok_or_else(|| anyhow!("missing markdown body in '{}'", path.display()))?;
-    let front_matter: IssuePromptFrontMatter =
-        serde_yaml::from_str(front_matter).with_context(|| {
-            format!(
-                "failed to parse issue prompt front matter '{}'",
-                path.display()
-            )
-        })?;
-    Ok(IssuePromptDoc {
-        front_matter,
-        body: body.trim_start().to_string(),
-    })
-}
-
-#[cfg(test)]
-fn resolve_issue_body(body: Option<String>, body_file: Option<&Path>) -> Result<String> {
-    if let Some(path) = body_file {
-        if path == Path::new("-") {
-            bail!("new: --body-file - is not supported in Rust path yet");
-        }
-        return fs::read_to_string(path)
-            .with_context(|| format!("new: --body-file not found: {}", path.display()));
-    }
-    Ok(body.unwrap_or_else(|| "## Goal\n-\n\n## Acceptance Criteria\n-".to_string()))
-}
-
-fn normalize_labels_csv(labels: &str, version: &str) -> String {
-    let mut normalized = labels
-        .split(',')
-        .map(str::trim)
-        .filter(|label| !label.is_empty() && !label.starts_with("version:"))
-        .map(str::to_string)
-        .collect::<Vec<_>>();
-    normalized.push(format!("version:{version}"));
-    normalized.join(",")
-}
-
-#[cfg(test)]
-fn parse_issue_number_from_url(url: &str) -> Result<u32> {
-    let issue = url
-        .trim()
-        .rsplit('/')
-        .next()
-        .ok_or_else(|| anyhow!("new: failed to parse issue number from URL: {url}"))?;
-    issue
-        .parse::<u32>()
-        .with_context(|| format!("new: failed to parse issue number from URL: {url}"))
 }
 
 fn path_relative_to_repo(repo_root: &Path, path: &Path) -> String {
