+++ adl/src/cli/pr_cmd_prompt.rs
@@ -1,0 +1,205 @@
+#[cfg(test)]
+use anyhow::{anyhow, Context};
+use anyhow::{bail, Result};
+#[cfg(test)]
+use serde::Deserialize;
+use std::fs;
+use std::path::{Path, PathBuf};
+
+use ::adl::control_plane::IssueRef;
+
+#[cfg(test)]
+#[derive(Debug, Deserialize)]
+pub(crate) struct IssuePromptFrontMatter {
+    pub(crate) title: String,
+    pub(crate) labels: Vec<String>,
+    pub(crate) issue_number: u32,
+}
+
+#[cfg(test)]
+#[derive(Debug)]
+pub(crate) struct IssuePromptDoc {
+    pub(crate) front_matter: IssuePromptFrontMatter,
+    pub(crate) body: String,
+}
+
+pub(crate) fn resolve_issue_scope_and_slug_from_local_state(
+    repo_root: &Path,
+    issue: u32,
+) -> Result<Option<(String, String)>> {
+    let issue_dir = format!("issue-{:04}__", issue);
+    let adl_root = repo_root.join(".adl");
+    if !adl_root.is_dir() {
+        return Ok(None);
+    }
+    let mut matches = Vec::new();
+    for scope_entry in fs::read_dir(&adl_root)? {
+        let scope_entry = scope_entry?;
+        let tasks = scope_entry.path().join("tasks");
+        if !tasks.is_dir() {
+            continue;
+        }
+        for entry in fs::read_dir(&tasks)? {
+            let entry = entry?;
+            let name = entry.file_name();
+            let name = name.to_string_lossy();
+            if let Some(slug) = name.strip_prefix(&issue_dir) {
+                matches.push((
+                    scope_entry.file_name().to_string_lossy().to_string(),
+                    slug.to_string(),
+                ));
+            }
+        }
+    }
+    matches.sort();
+    Ok(matches.into_iter().next())
+}
+
+pub(crate) fn render_generated_issue_prompt(
+    issue: u32,
+    slug: &str,
+    title: &str,
+    labels_csv: &str,
+    issue_url: &str,
+) -> String {
+    let wp = infer_wp_from_title(title);
+    let outcome_type = infer_required_outcome_type(labels_csv, title);
+    let label_lines = labels_csv
+        .split(',')
+        .map(str::trim)
+        .filter(|label| !label.is_empty())
+        .map(|label| format!("  - \"{label}\""))
+        .collect::<Vec<_>>()
+        .join("\n");
+
+    format!(
+        "---\nissue_card_schema: adl.issue.v1\nwp: \"{wp}\"\nslug: \"{slug}\"\ntitle: \"{title}\"\nlabels:\n{label_lines}\nissue_number: {issue}\nstatus: \"draft\"\naction: \"edit\"\ndepends_on: []\nmilestone_sprint: \"Pending sprint assignment\"\nrequired_outcome_type:\n  - \"{outcome_type}\"\nrepo_inputs: []\ncanonical_files: []\ndemo_required: false\ndemo_names: []\nissue_graph_notes:\n  - \"Bootstrap-generated from GitHub issue metadata because no canonical local issue prompt existed yet.\"\npr_start:\n  enabled: true\n  slug: \"{slug}\"\n---\n\n# {title}\n\n## Summary\n\nBootstrap-generated local source prompt for issue #{issue}.\n\n## Goal\n\nTranslate the GitHub issue into the canonical local STP/task-bundle flow and refine this prompt before execution as needed.\n\n## Required Outcome\n\nThis issue currently defaults to a required outcome type of `{outcome_type}`. Refine this if the issue requires a different outcome or a combination.\n\n## Deliverables\n\n- one bounded, reviewable outcome matching the issue scope\n- updated canonical docs/code/tests/demo artifacts as required by the issue\n\n## Acceptance Criteria\n\n- the issue title and labels are reflected in the local source prompt\n- the task can proceed through `pr init`, `pr start`, and card editing without manual bootstrap repair\n\n## Repo Inputs\n\n- {issue_url}\n\n## Dependencies\n\n- none recorded yet\n\n## Demo Expectations\n\n- No demo is required by default. Update this section if the issue requires a proof surface.\n\n## Non-goals\n\n- changing milestone scope without recording it explicitly\n- ad-hoc local workflow drift outside the tracked issue flow\n\n## Issue-Graph Notes\n\n- This prompt was generated automatically because the canonical local issue prompt was missing.\n- Review and refine it before substantive implementation work begins.\n\n## Notes\n\n- GitHub issue: {issue_url}\n\n## Tooling Notes\n\n- Generated by `pr.sh` bootstrap fallback.\n"
+    )
+}
+
+pub(crate) fn infer_wp_from_title(title: &str) -> String {
+    if let Some(start) = title.find("[WP-") {
+        if let Some(end_rel) = title[start + 1..].find(']') {
+            return title[start + 1..start + 1 + end_rel].to_string();
+        }
+    }
+    "unassigned".to_string()
+}
+
+pub(crate) fn infer_required_outcome_type(labels_csv: &str, title: &str) -> &'static str {
+    let lowered = format!("{} {}", labels_csv.to_lowercase(), title.to_lowercase());
+    if lowered.contains("type:docs")
+        || lowered.contains("area:docs")
+        || lowered.contains("[docs]")
+        || lowered.contains("type:design")
+    {
+        return "docs";
+    }
+    if lowered.contains("type:test") || lowered.contains("area:tests") || lowered.contains("[test]")
+    {
+        return "tests";
+    }
+    if lowered.contains("area:demo") || lowered.contains("[demo]") {
+        return "demo";
+    }
+    "code"
+}
+
+#[cfg(test)]
+pub(crate) fn version_from_labels_csv(labels_csv: &str) -> Option<String> {
+    labels_csv
+        .split(',')
+        .map(str::trim)
+        .find_map(|label| label.strip_prefix("version:").map(str::to_string))
+}
+
+pub(crate) fn version_from_title(title: &str) -> Option<String> {
+    let start = title.find("[v")?;
+    let rest = &title[start + 1..];
+    let end = rest.find(']')?;
+    Some(rest[..end].to_string())
+}
+
+pub(crate) fn validate_issue_prompt_exists(path: &Path) -> Result<()> {
+    if !path.is_file() {
+        bail!("missing canonical source issue prompt: {}", path.display());
+    }
+    Ok(())
+}
+
+pub(crate) fn resolve_issue_prompt_path(repo_root: &Path, issue_ref: &IssueRef) -> Result<PathBuf> {
+    let preferred = issue_ref.issue_prompt_path(repo_root);
+    if preferred.is_file() {
+        return Ok(preferred);
+    }
+
+    let legacy = issue_ref.legacy_issue_prompt_path(repo_root);
+    if legacy.is_file() {
+        return Ok(legacy);
+    }
+
+    bail!(
+        "missing canonical source issue prompt: {}",
+        preferred.display()
+    )
+}
+
+#[cfg(test)]
+pub(crate) fn load_issue_prompt(path: &Path) -> Result<IssuePromptDoc> {
+    let text = fs::read_to_string(path)
+        .with_context(|| format!("failed to read issue prompt '{}'", path.display()))?;
+    let mut parts = text.splitn(3, "---");
+    let _ = parts.next();
+    let front_matter = parts
+        .next()
+        .ok_or_else(|| anyhow!("missing front matter in '{}'", path.display()))?;
+    let body = parts
+        .next()
+        .ok_or_else(|| anyhow!("missing markdown body in '{}'", path.display()))?;
+    let front_matter: IssuePromptFrontMatter =
+        serde_yaml::from_str(front_matter).with_context(|| {
+            format!(
+                "failed to parse issue prompt front matter '{}'",
+                path.display()
+            )
+        })?;
+    Ok(IssuePromptDoc {
+        front_matter,
+        body: body.trim_start().to_string(),
+    })
+}
+
+#[cfg(test)]
+pub(crate) fn resolve_issue_body(body: Option<String>, body_file: Option<&Path>) -> Result<String> {
+    if let Some(path) = body_file {
+        if path == Path::new("-") {
+            bail!("new: --body-file - is not supported in Rust path yet");
+        }
+        return fs::read_to_string(path)
+            .with_context(|| format!("new: --body-file not found: {}", path.display()));
+    }
+    Ok(body.unwrap_or_else(|| "## Goal\n-\n\n## Acceptance Criteria\n-".to_string()))
+}
+
+pub(crate) fn normalize_labels_csv(labels: &str, version: &str) -> String {
+    let mut normalized = labels
+        .split(',')
+        .map(str::trim)
+        .filter(|label| !label.is_empty() && !label.starts_with("version:"))
+        .map(str::to_string)
+        .collect::<Vec<_>>();
+    normalized.push(format!("version:{version}"));
+    normalized.join(",")
+}
+
+#[cfg(test)]
+pub(crate) fn parse_issue_number_from_url(url: &str) -> Result<u32> {
+    let issue = url
+        .trim()
+        .rsplit('/')
+        .next()
+        .ok_or_else(|| anyhow!("new: failed to parse issue number from URL: {url}"))?;
+    issue
+        .parse::<u32>()
+        .with_context(|| format!("new: failed to parse issue number from URL: {url}"))
+}
